\name{specificities.distribution.plot}
\Rdversion{1.1}
\alias{specificities.distribution.plot}

\title{
  Display specificities probability
}

\description{
Display specificities probability distribution (call dhyper and \code{\link{specificities.probabilities.vector}})
}

\usage{
  specificities.distribution.plot(x, F, t, T)
}

\arguments{
  \item{x}{observed number of A words}
  \item{F}{total number of A}
  \item{t}{size of part}
  \item{T}{size of corpus}
}

\value{
  nothing
}

\author{ Matthieu Decorde, Serge Heiden }

