\name{specificities.lexicon.new}
\Rdversion{1.1}
\alias{specificities.lexicon.new}

\title{
specificities association score with two frequency list.
}

\description{
Compute specificities association score between a lexicon and a sub-lexicon.
A new version of the "specificities.lexicon" function
}

\usage{
specificities.lexicon.new(lexicon, sublexicon)
}

\arguments{
  \item{lexicon}{ a frequency list (named vector) }
  \item{sublexicon}{ a frequency list (named vector) }
}

\value{
  specificities index as a named vector.
}

\seealso{
  \code{\link{specificities}} for specificities score and references
}
