\name{specificities.lexicon.probabilities}
\Rdversion{1.1}
\alias{specificities.lexicon.probabilities}

\title{
  Raw hypergeometric probabilities
}
\description{
  Raw hypergeometric probabilities for computing specificities, see \code{\link{specificities.lexicon}}
}
\usage{
  specificities.lexicon.probabilities(lexicon, sublexicon)
}

\arguments{
  \item{lexicon}{See \code{\link{specificities.lexicon}}}
  \item{sublexicon}{See \code{\link{specificities.lexicon}}}
}

\value{
  Hypergeometric probabilities. See \code{\link{specificities.lexicon}}.
}

\author{ Sylvain Loiseau }

\seealso{
  See \code{\link{specificities.lexicon}}
}
