\name{specificities.lexicon}
\Rdversion{1.1}
\alias{specificities.lexicon}

\title{
*OBSOLETE FUNCTION (see 'specificities.lexicon.new')* specificities association score with two frequency lists.
}

\description{
Compute specificities association score between a lexicon and a sub-lexicon
}

\usage{
specificities.lexicon(lexicon, sublexicon)
}

\arguments{
  \item{lexicon}{ a frequency list (named vector) }
  \item{sublexicon}{ a frequency list (named vector) }
}

\value{
  specificities index as a named vector.
}

\seealso{
  \code{\link{specificities}} for specificities score and references
}
