\name{specificities.probabilities}
\Rdversion{1.1}
\alias{specificities.probabilities}

\title{
Raw hypergeometric probabilities
}
\description{
Raw hypergeometric probabilities, used for computing specificities (see \code{\link{specificities}}).
}

\usage{
specificities.probabilities(lexicaltable, types = NULL, parts = NULL)
}

\arguments{
  \item{lexicaltable}{ see \code{\link{specificities}} }

  \item{types}{ see \code{\link{specificities}} }

  \item{parts}{ see \code{\link{specificities}} }
}

\seealso{
 see \code{\link{specificities}}.
}
